requirements.txt
requests
websocket-client
regex